# docs/public/api/swagger-2.0.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/group/exports": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns export job rows for the caller's group, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "List Collection Exports",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Results-repo_ExportOut"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a pending export row and enqueues the build job. Poll the listing endpoint or watch the WebSocket for completion.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Start a Collection Export",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/repo.ExportOut"
                        }
                    }
                }
            }
        },
        "/v1/group/exports/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Get an Export",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Export ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.ExportOut"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes the export row and its blob artifact.",
                "tags": [
                    "Group"
                ],
                "summary": "Delete an Export",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Export ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/group/exports/{id}/download": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Download an Export Artifact",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Export ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/v1/group/import": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Uploads a collection-export zip and enqueues the import job. The destination group must be empty. Returns the tracked import row so clients can poll for progress.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Group"
                ],
                "summary": "Import a Collection Zip",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Export zip",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/repo.ExportOut"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "repo.ExportOut": {
            "type": "object",
            "properties": {
                "artifactPath": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "destinationId": {
                    "description": "DestinationID is the backup destination holding the artifact; absent\nfor plain manual exports, which live in the primary storage.",
                    "type": "string",
                    "x-nullable": true
                },
                "error": {
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is \"export\" for server-produced backup artifacts, \"import\" for\nuser-uploaded restore zips. The lifecycle fields below behave the\nsame for both.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin is \"manual\" for user-triggered exports and \"scheduled\" for\nbackups started by the scheduler.",
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "v1.Results-repo_ExportOut": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ExportOut"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}